finished services

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -5,24 +5,24 @@ package bgu.spl.mics.application.objects;
  * These points are used to generate a point cloud representing objects in the environment.
  */
 public class CloudPoint {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     
-    public CloudPoint(int x, int y)
+    public CloudPoint(double x, double y)
     {
         this.x = x;
         this.y = y;
     }
-    public void setPos(int x, int y)
+    public void setPos(double x, double y)
     {
         this.x = x;
         this.y = y;
     }
-    public int getX()
+    public double getX()
     {
         return this.x;
     }
-    public int getY()
+    public double getY()
     {
         return this.y;
     }
